backend/requirements.txt
# FastAPI相关依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 异步任务处理
celery==5.3.4
redis==5.0.1

# PPT处理
python-pptx==0.6.23

# 环境配置
python-dotenv==1.0.0
pydantic-settings==2.1.0

# 数据验证
pydantic==2.5.0

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.13.2